refactor: replace custom badge via harmony Badge
