+++ src/components/UserBadge/UserBadge.module.css
@@ -1,0 +1,3 @@
+.UserBadge {
+    color: var(--text-secondary);
+}
